legacy: move existing navigator to legacy
